move server code out of codegen dir
