+++ _tasks/gen_deploy.ts
@@ -1,5 +1,5 @@
-import { getModuleIndex, getSha } from "../codegen/server/git_utils.ts"
 import { ensureDir } from "../deps/std/fs.ts"
+import { getModuleIndex, getSha } from "../server/git_utils.ts"
 
 const sha = await getSha()
 const index = await getModuleIndex()
