+++ cache.ts
@@ -1,5 +1,5 @@
-import { LocalCapiCodegenServer } from "./codegen/server/local.ts"
 import * as fs from "./deps/std/fs.ts"
+import { LocalCapiCodegenServer } from "./server/local.ts"
 
 await fs.emptyDir("target/codegen/generated")
 const port = 5646
