+++ serve.ts
@@ -1,4 +1,4 @@
-import { LocalCapiCodegenServer } from "./codegen/server/local.ts"
+import { LocalCapiCodegenServer } from "./server/local.ts"
 
 console.log("Listening on http://localhost:5646/")
 new LocalCapiCodegenServer().listen(5646)
