+++ server/capi_repo.ts
@@ -1,4 +1,4 @@
-import { PermanentMemo, TimedMemo } from "../../util/memo.ts"
+import { PermanentMemo, TimedMemo } from "../util/memo.ts"
 import { getFullSha, getSha, SHA_ABBREV_LENGTH } from "./git_utils.ts"
 import { CodegenServer } from "./server.ts"
