+++ server/codecs.test.ts
@@ -2,9 +2,9 @@ import { Codec } from "../deps/scale.ts"
 import { assertEquals } from "../deps/std/testing/asserts.ts"
 import { DeriveCodec } from "../scale_info/mod.ts"
 import * as testClients from "../test_util/clients/mod.ts"
-import { InMemoryCache } from "./server/cache.ts"
-import { LocalCapiCodegenServer } from "./server/local.ts"
-import { highlighterPromise } from "./server/server.ts"
+import { InMemoryCache } from "./cache.ts"
+import { LocalCapiCodegenServer } from "./local.ts"
+import { highlighterPromise } from "./server.ts"
 
 await highlighterPromise
