+++ server/server.ts
@@ -1,13 +1,13 @@
 import { serve } from "https://deno.land/std@0.165.0/http/server.ts"
 import { escapeHtml } from "https://deno.land/x/escape@1.4.2/mod.ts"
 import * as shiki from "https://esm.sh/shiki@0.11.1?bundle"
-import * as $ from "../../deps/scale.ts"
-import * as C from "../../mod.ts"
-import * as T from "../../test_util/mod.ts"
-import * as U from "../../util/mod.ts"
-import { TimedMemo } from "../../util/mod.ts"
-import { Files } from "../Files.ts"
-import { codegen } from "../mod.ts"
+import { Files } from "../codegen/Files.ts"
+import { codegen } from "../codegen/mod.ts"
+import * as $ from "../deps/scale.ts"
+import * as C from "../mod.ts"
+import * as T from "../test_util/mod.ts"
+import * as U from "../util/mod.ts"
+import { TimedMemo } from "../util/mod.ts"
 import { Cache } from "./cache.ts"
 
 shiki.setCDN("https://unpkg.com/shiki/")
